proyecto 2/requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.3
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
reportlab>=4.0.0
werkzeug>=3.0.0
gunicorn>=22.0.0
